pyproject.toml
[project]
authors = [
    {name = "Henry Zhu", email = "daya0576@gmail.com"},
]
name = "beaverhabits"
version = "0.4.4"
requires-python = "<4.0,>=3.12"
dependencies = [
    "nicegui==2.22.2",
    "pydantic-settings<3.0.0,>=2.2.1",
    "fastapi-users[sqlalchemy]<15.0.0,>=14.0.0",
    "uvicorn[standard]<1.0.0,>=0.28.0",
    "sqlalchemy[asyncio]<3.0.0,>=2.0.29",
    "asyncpg<1.0.0,>=0.29.0",
    "aiosqlite<1.0.0,>=0.20.0",
    "pytz<2025.0,>=2024.1",
    "sentry-sdk[fastapi]<3.0.0,>=2.15.0",
    "python-dateutil<3.0.0.0,>=2.9.0.post0",
    "loguru>=0.7.3",
    "cachetools>=5.5.1",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.24.0",
]
description = ""
readme = "README.md"

[dependency-groups]
dev = [
    "icecream<3.0.0,>=2.1.3",
    "autopep8<3.0.0,>=2.0.4",
    "pytest-selenium<5.0.0,>=4.1.0",
    "pytest-asyncio<1.0.0.0,>=0.23.5.post1",
    "webdriver-manager<5.0.0,>=4.0.1",
    "selenium<5.0.0,>=4.18.1",
    "beautifulsoup4<5.0.0,>=4.12.3",
    "pytest<9.0.0,>=8.1.1",
    "pip>=24.3.1",
    "psutil>=6.1.1",
]

[project.scripts]
main = "beaverhabits.routes:main"

